+++ messages/main.go
@@ -4,19 +4,19 @@ import (
 	"github.com/micro/micro/v3/service"
 	"github.com/micro/micro/v3/service/logger"
 
-	"github.com/micro/services/chat/handler"
-	pb "github.com/micro/services/chat/proto"
+	"github.com/micro/services/messages/handler"
+	pb "github.com/micro/services/messages/proto"
 )
 
 func main() {
 	// Create the service
 	srv := service.New(
-		service.Name("chat"),
+		service.Name("messages"),
 		service.Version("latest"),
 	)
 
 	// Register the handler against the server
-	pb.RegisterChatHandler(srv.Server(), new(handler.Chat))
+	pb.RegisterMessagesHandler(srv.Server(), new(handler.Messages))
 
 	// Run the service
 	if err := srv.Run(); err != nil {
